Changed floats to doubles

+++ src/pl/sda/Main.java
@@ -6,15 +6,15 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("What is your weight in kg?");
-        float weight = sc.nextFloat();
+        double weight = sc.nextDouble();
         System.out.println("What is your height in m?");
-        float height = sc.nextFloat();
-        float bmi=bmi(weight, height);
+        double height = sc.nextDouble();
+        double bmi=bmi(weight, height);
         System.out.println(bmi);
 
     }
 
-    private static float bmi(float weight, float height) {
+    private static double bmi(double weight, double height) {
         return weight/(height*height);
     }
 }
